Add check if user is an author of comment

+++ src/features/comments/controllers/deleteCommentController.ts
@@ -1,14 +1,20 @@
 import { Request, Response } from 'express';
 import { ErrorResultModel } from '../../../models/ErrorResultModel';
 import { commentsService } from '../../../domains/comments-service';
+import { commentsDBRepository } from '../../../repositories/commentsDBRepository';
 
 
-export const deleteCommentController = async (req: Request<{commentId: string}>, res: Response<null | ErrorResultModel>) => {
-  const commentToDelete = await commentsService.deleteComment(req.params.commentId)
-  if (!commentToDelete) {
-      res.status(404).json({ errorsMessages: [{ message: 'Commentr not found', field: 'commentId' }] })
-      return
-  }
-  res
-      .sendStatus(204)
+export const deleteCommentController = async (req: Request<{ commentId: string }>, res: Response<null | ErrorResultModel>) => {
+    const isUserAuthorOfComment = await commentsDBRepository.checkIfUserIfAuthorOfComment(req.user!, req.params.commentId)
+    if (!isUserAuthorOfComment) {
+        res.status(403).json({ errorsMessages: [{ message: 'You are not the author of this comment', field: 'userId' }] })
+    } else {
+        const commentToDelete = await commentsService.deleteComment(req.params.commentId)
+        if (!commentToDelete) {
+            res.status(404).json({ errorsMessages: [{ message: 'Comment not found', field: 'commentId' }] })
+            return
+        }
+        res
+            .sendStatus(204)
+    }
 };
